requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.5
django-crispy-forms==1.14.0
django-csp==3.7
django-js-asset==2.0.0
django-permissions-policy==4.14.0
django-tinymce==3.5.0
Pillow==9.4.0
sqlparse==0.4.3
tzdata==2022.7
